feat(funcions): Adicionado a função title

+++ main.py
@@ -7,15 +7,21 @@ while True:
   print('[3] Lookup a device')
   print('[4] Delete a device')
   print('[5] Quit')
-  option = int(input('Option >> '))
-  if option == 1:
-    functions.Create()
-  elif option == 2:
-    functions.GetAll()
-  elif option == 3:
-    functions.GetDevice()
-  elif option == 4:
-    functions.deleteByLine()
+  option = input('Option >> ')
+  if option.isdigit():
+    option = int(option)
+    if option == 1:
+      functions.Create()
+    elif option == 2:
+      functions.GetAll()
+    elif option == 3:
+      functions.GetDevice()
+    elif option == 4:
+      functions.deleteByLine()
+    elif option == 5:
+      print('Program terminated')
+      break
+    else:
+      print('Invalid option!')
   else:
-    print('Program terminated')
-    break
+    print('invalid option!')
